+++ src/index.ts
@@ -1,4 +1,5 @@
 export { up } from './up'
 export type { FetcherOptions, DefaultOptions, ResolvedOptions } from './types'
-export { type ResponseError, isResponseError } from './response-error'
+export { ResponseError, isResponseError } from './response-error'
 export { type ValidationError, isValidationError } from './validation-error'
+export { isJsonifiable } from './utils'