+++ src/up.ts
@@ -1,37 +1,44 @@
 import type { StandardSchemaV1 } from '@standard-schema/spec'
 import { resolveOptions } from './resolve-options'
-import type { FetcherOptions, DefaultOptions, BaseFetchFn } from './types'
+import type {
+   FetcherOptions,
+   DefaultOptions,
+   BaseFetchFn,
+   FallbackOptions,
+} from './types'
 import { emptyOptions, parseStandardSchema } from './utils'
 
 export function up<
    TFetchFn extends BaseFetchFn,
-   TDefaultOptions extends DefaultOptions<TFetchFn> = DefaultOptions<TFetchFn>,
+   TDefaultParsedData = any,
+   TDefaultRawBody = Parameters<FallbackOptions<any>['serializeBody']>[0],
 >(
    fetchFn: TFetchFn,
-   getDefaultOptions: () => TDefaultOptions = () => emptyOptions,
+   getDefaultOptions: (
+      input: Parameters<TFetchFn>[0],
+      fetcherOpts: FetcherOptions<TFetchFn, any, any, any>,
+      ctx?: Parameters<TFetchFn>[2],
+   ) => DefaultOptions<TFetchFn, TDefaultParsedData, TDefaultRawBody> = () =>
+      emptyOptions,
 ) {
    return <
-      TParsedData = Awaited<
-         ReturnType<NonNullable<TDefaultOptions['parseResponse']>>
-      >,
+      TParsedData = TDefaultParsedData,
       TSchema extends StandardSchemaV1<
          TParsedData,
          any
       > = StandardSchemaV1<TParsedData>,
+      TRawBody = TDefaultRawBody,
    >(
       input: Parameters<TFetchFn>[0],
-      fetcherOptions:
-         | FetcherOptions<TFetchFn, TSchema, TParsedData>
-         | ((
-              defaultOptions: TDefaultOptions,
-           ) => FetcherOptions<TFetchFn, TSchema, TParsedData>) = emptyOptions,
+      fetcherOpts: FetcherOptions<
+         TFetchFn,
+         TSchema,
+         TParsedData,
+         TRawBody
+      > = emptyOptions,
       ctx?: Parameters<TFetchFn>[2],
    ) => {
-      let defaultOpts = getDefaultOptions()
-      let fetcherOpts =
-         typeof fetcherOptions === 'function'
-            ? fetcherOptions(defaultOpts)
-            : fetcherOptions
+      let defaultOpts = getDefaultOptions(input, fetcherOpts, ctx)
       let options = resolveOptions(input, defaultOpts, fetcherOpts)
       defaultOpts.onRequest?.(options)